+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+* Add `Translator.createGlossaryFromCsvAsync()` allowing glossaries downloaded
+  from website to be easily uploaded to API.
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+
 ## [1.2.1] - 2022-06-29
 ### Changed
 * Update contributing guidelines, we can now accept Pull Requests.
@@ -78,6 +89,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.2.1..HEAD
 [1.2.1]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.2.0..v1.2.1
 [1.2.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.1.0..v1.2.0
 [1.1.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.0.5..v1.1.0