+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Added
+* New language available: Ukrainian (`'uk'`). Add language code constant and
+  tests.
+
+  Note: older library versions also support new languages, this update only
+  adds new code constant.
+
+
 ## [1.3.0] - 2022-08-02
 ### Added
 * Add `Translator.createGlossaryFromCsvAsync()` allowing glossaries downloaded
@@ -86,6 +95,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.3.0..HEAD
 [1.3.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.2.1..v1.3.0
 [1.2.1]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.2.0..v1.2.1
 [1.2.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.1.0..v1.2.0