+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Fixed
+* Fix CI build for netcoreapp3.1.
+
+
 ## [1.5.0] - 2022-09-30
 ### Added
 * Add new `Formality` options: `PreferLess` and `PreferMore`.
@@ -104,6 +109,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.5.0...HEAD
 [1.5.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.2.1...v1.3.0