+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [1.5.1] - 2023-01-25
 ### Fixed
 * Fix CI build for netcoreapp3.1.
 * Fix `Formality` options `PreferLess` and `PreferMore` for document
@@ -111,7 +111,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release.
 
 
-[Unreleased]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.5.0...HEAD
+[1.5.1]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/DeepLcom/deepl-dotnet/compare/v1.2.1...v1.3.0